Rename new source files. #2
